Removing unused display_order reference from PaymentForm
The reference used to be there and managed order creation once the
payment was created. This stopped working after payment items refactoring
and the references remained there.

For the sake of opensourcing the module, let's remove the references
and whole order handling will have to be handled in the future if it's
still desirable.

remp/crm#647