Making sure that payment's gateway is usable in PaymentFormFactory
When editing the payment which uses gateway that's not registered
anymore, we need to make sure that unregistered gateway is allowed
to be used in the gateway select box so the payment can be still
updated.

Once the non-registered gateway is not used at the payment, it's not
offered to be used anymore.

remp/crm#714