Bugfixing CSOB mail parser again, raising same payment limit in processor
The 5 days limit in mail processor was enough, but caused problems
when we needed to process emails for the couple of last days.
If there was a weekend/holidays, the "created_at" field payment could
have been older than 5 days and helpdesk could have confirmed the
payment manually in the meantime.

This should give us more headroom without breaking anything.