Handle cancelled/error states in payment processors only for FAIL payments
At this point the error/cancelled states were handled also for payments
which could have been in FORM state, because their complete() method
returned `null` to indicate that payment was not failed nor successful.

Some gateways can implement "not settled" state to indicate, that the payment
will be confirmed soon, but some (such as bank transfer) do not.
For these kind of gateways we need to make sure they'll be handled by FORM
flow in the payment processor callback.

This also fixes the issue with duplicate payment logs for same payment
if it ended up with error.

remp/crm-payments-module#6