Fix recurrent payment reactivation
The status was changed correctly, but the recurrent payment
still linked to the failed attempt and was never attempted
again.

It's necessary to get rid of link to the failed payment
and also to clear approval/status fields just to be sure
there are no obsolete data there for the next attempt.

remp/crm#1688