Move payment gateway before variable symbol in payment form
Variable symbol generation has been recently changed to accept
the payment gateway as an argument. That means that it can
vary based on the selected gateway and logically it should
be the gateway which is selected first.

Any further change on the gateway will clear generated variable
symbol and will require new one to be requested.

remp/crm#1440