Fix bug in the payment editation form if pre-authorized payment is being edited
The "authorized" state is not among other allowed states within the form.
If someone tried to open such payment, it would fail on invalid default
value within the status field.

Since we don't want to have "authorized" payments created via admin, we only
add this option when payment is edited and "authorized" is already selected.