Use custom fork of openpgp-php to handle PHP 8.1 notices
We temporarily replace singpolyma/openpgp-php with the fork of this
library in order to prevent deprecation notices in PHP 8.1.

These notices were already handled in the library and were part of the
0.6 release. However one of the dependencies (phpseclib/phpseclib) is
locked to the v2.* by google/auth library and the 0.6 release requires
v3.* of this library and we're not able to upgrade it at the moment.

Once there's a chance to use phpseclib v3, we can revert this commit.

remp/crm#2706