Fix transitive openpgp-php dependency
The original idea was to use custom fork and define it in composer
as the repository pointing to the specific commit. This worked for
us but didn't work after releasing the open-source modules.

We made the composer package from the forked version so it can be
fetched correctly.

Full explanation why we do this is in the original commit:
e1c409be9ec7221f8c3130ea044385b921d27cc5

remp/crm#2773

(cherry picked from commit 326ce25b9f82ebfc8621eae3144396fdb3a644e4)