Fix duplication of VS for immutable copies
If the payments was created as a copy of immutable
payment and we only changed the gateway, the variable
symbol of the newly created payment didn't change.

This caused issues with confirming the payments, when
we selected incorrect instance based on the variable
symbol and incorrectly sent the gateway return process
to the "failed" scenario.

remp/helpdesk#2613