Always recalculate widget numbers when recalculating segments
This is "bugfix". But we added option `--no_widgets` to
`segment:actualize_counts` in case you want to keep previous behavior.
If option is used, command will not recalculate statistics for widgets
visible on the segment detail screen, only the number of people within
the segment.

Also added tooltip which explains average payment amount in segment detail.
Calculated average payment amount in segment detail is not ARPU. It's
average of "monthly averages" of all users in segment.

Co-authored-by: Peter Dulacka <dulacka@gmail.com>

remp/crm#3193