Fix possible type error when editing payment
The error appeared for payments with items without
link to subscription type item. JS changed these to empty
strings and they were not compliant with ?int type
on the backend.

If there's an empty string as the value of the hidden
element, we convert it to null right in the JS so the
frontend submits it the same way as it received it
from backend.

remp/respekt#317