Add custom subscription renewals
- Add `respekt:subscriptions_category` command to assign subscriptions
to subscription category.

- Add option to set renewal category for subscription.

- Add widget to show subscription current subscription category 
and renewal category if set.

- Add new scenario generic event `RespektAttachRenewalPaymentEvent` 
to create and attach renewal payment to subscription in scenario.

- Add scenario criteria `HasCustomRenewalCategoryCriteria` to check 
if subscription has renewal payment already attached to subscription 
or custom renewal category is set.

- Add scenario criteria `IsRenewalPaymentCategoryCriteria` to check 
renewal category of renewal payment. When renewal payment is paid, 
subscription with category od renewal payment is created.

remp/respekt#332