Add proforma invoice generation to all renewal payments
Until now, only the custom scenario generic action attached
proforma invoice to the renewal payment emails. This has
been refactored so that every email is checked and assigned
proforma invoice if necessary.

The invoice is attached only to the emails with renewal
payment parameter, which is assigned to a subscription and
it wasn't paid yet.

remp/respekt#408